+++ src/mass_driver/drivers/__init__.py
@@ -1,4 +1,5 @@
 """All the Patch Drivers we've implemented"""
 from mass_driver.drivers.counter import Counter
+from mass_driver.drivers.poetry import Poetry
 
-__all__ = ["Counter"]
+__all__ = ["Counter", "Poetry"]