+++ src/mass_driver/main.py
@@ -3,7 +3,7 @@
 from pathlib import Path
 from typing import Optional
 
-from mass_driver.drivers import Counter
+from mass_driver.drivers import Poetry
 from mass_driver.forges import GithubForge
 from mass_driver.model import Forge, PatchDriver
 from mass_driver.repo import clone_if_remote, commit
@@ -13,7 +13,7 @@ def main(
     repo_paths: list[str], do_patch: bool, branch_name: Optional[str], auth_token: str
 ):
     """Run the program's main command"""
-    driver = Counter(counter_file=Path("counter"), target_count=1)
+    driver = Poetry(package="pytest", target_major="8")
     forge = GithubForge(auth_token)
     if not branch_name:
         branch_name = driver.__class__.__name__.lower()
@@ -34,16 +34,17 @@ def process_repo(
 ):
     """Process a repo with Mass Driver: detect or patch"""
     repo = clone_if_remote(repo_path)
+    repo_as_path = Path(repo.working_dir)
     if do_patch:
         print(f"Detecting '{repo_path}' before patching...")
-        needs_patch = driver.detect(repo.working_dir)
+        needs_patch = driver.detect(repo_as_path)
         if not needs_patch:
             print("Does not need patching: skipping")
             return
-        driver.patch(repo.working_dir)
+        driver.patch(repo_as_path)
         print("Done patching, committing")
         commit(repo, driver, branch_name)
         print("Done committing")
     else:
-        needs_patch = driver.detect(repo.working_dir)
+        needs_patch = driver.detect(repo_as_path)
         print(f"Detecting '{repo_path}': {needs_patch=}")